HS-83 FEAT: 1차 구현
