+++ src/utils/replace_dollar/replace_dollar.c
@@ -1,23 +1,5 @@
 #include "../../../include/minishell.h"
 
-static int	get_env_len(char *str)
-{
-	int	len;
-
-	len = 0;
-	if (!str)
-		return (0);
-	while (*str)
-	{
-		if (check_white_space(*str) || *str == '\'' || *str == '\"' || *str == '$'
-			|| *str == '|' || *str == '=' || *str == '/' || *str == '\\')
-			break ;
-		len++;
-		str++;
-	}
-	return (len);
-}
-
 static char	*replace_dollar(char *input_buffer, char *temp, t_minishell *minishell)
 {
 	char	*replaced_value;
@@ -45,97 +27,29 @@ static char	*replace_dollar(char *input_buffer, char *temp, t_minishell *minishe
 	return (new_input_buffer);
 }
 
-char	*append_buffer_under_dollar(char *save, char const *buffer)
+static void	handle_single_quote(char **input_buffer, char **input_ptr)
 {
-	char	*new;
-	char	*temp;
-	char	origin_value;
-
-	temp = ft_strchr(buffer, '$');
-	if (buffer == NULL || temp == NULL)
-		return (save);
-	origin_value = *temp;
-	*temp = '\0';
-	if (save == NULL && buffer)
+	if (ft_strchr(*input_ptr, '\'') < ft_strchr(*input_ptr, '$'))
 	{
-		new = safe_malloc(ft_strlen(buffer) + 1);
-		ft_strlcpy(new, buffer, ft_strlen(buffer) + 1);
-		return (new);
+		*input_buffer = append_buffer_under_single_quote(*input_buffer, *input_ptr);
+		*input_ptr += get_under_single_quote_len(*input_ptr);
+		*input_buffer = append_single_quote(*input_buffer, *input_ptr, get_single_quote_len(*input_ptr));
+		if (get_single_quote_len(*input_ptr) > 0)
+			*input_ptr += get_single_quote_len(*input_ptr) + 2;
 	}
-	new = safe_malloc(ft_strlen(save) + ft_strlen(buffer) + 1);
-	ft_strlcpy(new, save, ft_strlen(save) + 1);
-	ft_strlcpy(new + ft_strlen(save), buffer, ft_strlen(buffer) + 1);
-	free(save);
-	save = NULL;
-	*temp = origin_value;
-	return (new);
-}
-char	*append_buffer_under_single_quote(char *save, char const *buffer)
-{
-	char	*new;
-	char	*temp;
-	char	origin_value;
-
-	temp = ft_strchr(buffer, '\'');
-	if (buffer == NULL || temp == NULL)
-		return (save);
-	origin_value = *temp;
-	*temp = '\0';
-	if (save == NULL && buffer)
-	{
-		new = safe_malloc(ft_strlen(buffer) + 1);
-		ft_strlcpy(new, buffer, ft_strlen(buffer) + 1);
-		return (new);
-	}
-	new = safe_malloc(ft_strlen(save) + ft_strlen(buffer) + 1);
-	ft_strlcpy(new, save, ft_strlen(save) + 1);
-	ft_strlcpy(new + ft_strlen(save), buffer, ft_strlen(buffer) + 1);
-	free(save);
-	save = NULL;
-	*temp = origin_value;
-	return (new);
-}
-
-char	*append_single_quote(char *input_buffer, char *input_ptr, int single_quote_len)
-{
-	char	*new_input_buffer;
-
-	if (single_quote_len <= 0)
-		return ;
-	new_input_buffer = safe_malloc(ft_strlen(input_buffer) + single_quote_len + 1);
-	ft_strlcpy(new_input_buffer, input_buffer, ft_strlen(input_buffer) + 1);
-	fr_strlcpy(new_input_buffer + ft_strlen(input_buffer), input_ptr, single_quote_len + 1);
-	free(input_buffer);
-	input_buffer = NULL;
-	return (new_input_buffer);
-}
-
-static int	get_single_quote_len(char *input_ptr)
-{
-	int	i;
-
-	i = ft_strchr(input_ptr + 1, '\'') - ft_strchr(input_ptr, '\'') - 1;
-	return (i);
 }
 
 char	*replace_whole_input_dollar(char *input, t_minishell *minishell)
 {
 	char	*input_buffer;
 	char	*input_ptr;
-	int		single_quote_len;
 	
 	if (!ft_strchr(input, '$'))
 		return (input);
 	input_buffer = safe_malloc(ft_strlen(input));
 	input_ptr = input;
-	single_quote_len = get_single_quote_len(input_ptr);
-	if (ft_strchr(input_ptr, '\'') < ft_strchr(input_ptr, '$'))
-	{
-		input_buffer = append_buffer_under_single_quote(input_buffer, input_ptr);
-		input_ptr += (ft_strchr(input_ptr, '\'') - input_ptr);
-		input_buffer = append_single_quote(input_buffer, input_ptr, get_single_quote_len(input_ptr));
-	}
-	ft_memccpy_under(input_buffer + single_quote_len, input_ptr, '$', ft_strlen(input_ptr));
+	handle_single_quote(&input_buffer, &input_ptr);
+	ft_memccpy_under(input_buffer, input_ptr, '$', ft_strlen(input_ptr));
 	while (TRUE)
 	{
 		input_ptr = ft_strchr(input_ptr, '$');
@@ -148,12 +62,7 @@ char	*replace_whole_input_dollar(char *input, t_minishell *minishell)
 		}
 		if (!input_ptr)
 			break ;
-		if (ft_strchr(input_ptr, '\'') < ft_strchr(input_ptr, '$'))
-		{
-			input_buffer = append_buffer_under_single_quote(input_buffer, input_ptr);
-			input_ptr = (ft_strchr(input_ptr, '\'') - input_ptr);
-			handle_single_quote(&input_buffer, &input_ptr, get_single_quote_len(input_ptr));
-		}
+		handle_single_quote(&input_buffer, &input_ptr);
 		if (ft_strchr(input_ptr, '$'))
 		{
 			input_buffer = append_buffer_under_dollar(input_buffer, input_ptr);
